Revert "Change map "
