--- data/processed/map_df.csv
@@ -1,111 +1,0 @@
-name,year,count,lat,lon
-Arbutus Ridge,2018,353,49.24680491,-123.1616692
-Arbutus Ridge,2019,345,49.24680491,-123.1616692
-Arbutus Ridge,2017,425,49.24680491,-123.1616692
-Arbutus Ridge,2021,273,49.24680491,-123.1616692
-Arbutus Ridge,2020,353,49.24680491,-123.1616692
-Central Business District,2019,14761,49.28074707,-123.116567
-Central Business District,2017,12409,49.28074707,-123.116567
-Central Business District,2020,9094,49.28074707,-123.116567
-Central Business District,2021,7833,49.28074707,-123.116567
-Central Business District,2018,13312,49.28074707,-123.116567
-Dunbar-Southlands,2018,502,49.23796191,-123.1895468
-Dunbar-Southlands,2020,488,49.23796191,-123.1895468
-Dunbar-Southlands,2019,470,49.23796191,-123.1895468
-Dunbar-Southlands,2021,305,49.23796191,-123.1895468
-Dunbar-Southlands,2017,498,49.23796191,-123.1895468
-Fairview,2018,2151,49.26454049,-123.1310489
-Fairview,2019,2594,49.26454049,-123.1310489
-Fairview,2021,1486,49.26454049,-123.1310489
-Fairview,2017,2086,49.26454049,-123.1310489
-Fairview,2020,2041,49.26454049,-123.1310489
-Grandview-Woodland,2017,2212,49.27643961,-123.0667282
-Grandview-Woodland,2018,2011,49.27643961,-123.0667282
-Grandview-Woodland,2021,1378,49.27643961,-123.0667282
-Grandview-Woodland,2020,1691,49.27643961,-123.0667282
-Grandview-Woodland,2019,2135,49.27643961,-123.0667282
-Hastings-Sunrise,2020,1185,49.27793405,-123.0402699
-Hastings-Sunrise,2018,1414,49.27793405,-123.0402699
-Hastings-Sunrise,2019,1533,49.27793405,-123.0402699
-Hastings-Sunrise,2017,1351,49.27793405,-123.0402699
-Hastings-Sunrise,2021,1078,49.27793405,-123.0402699
-Kensington-Cedar Cottage,2017,1571,49.24668645,-123.0728846
-Kensington-Cedar Cottage,2021,1332,49.24668645,-123.0728846
-Kensington-Cedar Cottage,2019,1807,49.24668645,-123.0728846
-Kensington-Cedar Cottage,2018,1588,49.24668645,-123.0728846
-Kensington-Cedar Cottage,2020,1533,49.24668645,-123.0728846
-Kerrisdale,2017,460,49.22365549,-123.1595762
-Kerrisdale,2020,446,49.22365549,-123.1595762
-Kerrisdale,2018,482,49.22365549,-123.1595762
-Kerrisdale,2019,511,49.22365549,-123.1595762
-Kerrisdale,2021,413,49.22365549,-123.1595762
-Killarney,2019,678,49.21702167,-123.037647
-Killarney,2018,646,49.21702167,-123.037647
-Killarney,2021,466,49.21702167,-123.037647
-Killarney,2017,658,49.21702167,-123.037647
-Killarney,2020,534,49.21702167,-123.037647
-Kitsilano,2017,1794,49.26753985,-123.1632947
-Kitsilano,2019,1853,49.26753985,-123.1632947
-Kitsilano,2021,1288,49.26753985,-123.1632947
-Kitsilano,2018,2045,49.26753985,-123.1632947
-Kitsilano,2020,1734,49.26753985,-123.1632947
-Marpole,2021,849,49.21020747,-123.1283817
-Marpole,2020,874,49.21020747,-123.1283817
-Marpole,2017,975,49.21020747,-123.1283817
-Marpole,2019,1048,49.21020747,-123.1283817
-Marpole,2018,959,49.21020747,-123.1283817
-Mount Pleasant,2018,2769,49.26306525,-123.0985131
-Mount Pleasant,2017,2710,49.26306525,-123.0985131
-Mount Pleasant,2019,3023,49.26306525,-123.0985131
-Mount Pleasant,2021,2074,49.26306525,-123.0985131
-Mount Pleasant,2020,2305,49.26306525,-123.0985131
-Oakridge,2019,670,49.2264032,-123.1230246
-Oakridge,2020,391,49.2264032,-123.1230246
-Oakridge,2017,615,49.2264032,-123.1230246
-Oakridge,2021,330,49.2264032,-123.1230246
-Oakridge,2018,589,49.2264032,-123.1230246
-Renfrew-Collingwood,2020,2052,49.24734273,-123.0401665
-Renfrew-Collingwood,2019,2030,49.24734273,-123.0401665
-Renfrew-Collingwood,2017,2219,49.24734273,-123.0401665
-Renfrew-Collingwood,2018,1987,49.24734273,-123.0401665
-Renfrew-Collingwood,2021,1892,49.24734273,-123.0401665
-Riley Park,2017,973,49.24476648,-123.1031468
-Riley Park,2020,840,49.24476648,-123.1031468
-Riley Park,2018,1015,49.24476648,-123.1031468
-Riley Park,2021,701,49.24476648,-123.1031468
-Riley Park,2019,1164,49.24476648,-123.1031468
-Shaughnessy,2020,412,49.24568098,-123.1397602
-Shaughnessy,2018,460,49.24568098,-123.1397602
-Shaughnessy,2017,481,49.24568098,-123.1397602
-Shaughnessy,2021,304,49.24568098,-123.1397602
-Shaughnessy,2019,367,49.24568098,-123.1397602
-South Cambie,2017,385,49.24555646,-123.1218009
-South Cambie,2018,346,49.24555646,-123.1218009
-South Cambie,2021,338,49.24555646,-123.1218009
-South Cambie,2019,482,49.24555646,-123.1218009
-South Cambie,2020,460,49.24555646,-123.1218009
-Strathcona,2017,2422,49.27821958,-123.0882353
-Strathcona,2019,2884,49.27821958,-123.0882353
-Strathcona,2020,2514,49.27821958,-123.0882353
-Strathcona,2018,2647,49.27821958,-123.0882353
-Strathcona,2021,1810,49.27821958,-123.0882353
-Sunset,2019,1210,49.21875615,-123.092038
-Sunset,2020,998,49.21875615,-123.092038
-Sunset,2017,938,49.21875615,-123.092038
-Sunset,2021,860,49.21875615,-123.092038
-Sunset,2018,1099,49.21875615,-123.092038
-Victoria-Fraserview,2020,587,49.22001194,-123.0641347
-Victoria-Fraserview,2017,752,49.22001194,-123.0641347
-Victoria-Fraserview,2018,686,49.22001194,-123.0641347
-Victoria-Fraserview,2021,541,49.22001194,-123.0641347
-Victoria-Fraserview,2019,710,49.22001194,-123.0641347
-West End,2020,2659,49.28501119,-123.1354377
-West End,2019,3817,49.28501119,-123.1354377
-West End,2018,3526,49.28501119,-123.1354377
-West End,2017,3422,49.28501119,-123.1354377
-West End,2021,2284,49.28501119,-123.1354377
-West Point Grey,2019,435,49.26840121,-123.2034675
-West Point Grey,2018,408,49.26840121,-123.2034675
-West Point Grey,2017,444,49.26840121,-123.2034675
-West Point Grey,2021,328,49.26840121,-123.2034675
-West Point Grey,2020,454,49.26840121,-123.2034675
