--- src/app_original.py
@@ -1,318 +1,0 @@
-from dash import Dash, html, Input, Output, dcc
-import dash_bootstrap_components as dbc
-import pandas as pd
-import numpy as np
-import altair as alt
-import plotly.express as px
-import geopandas as gpd
-
-alt.data_transformers.disable_max_rows()
-
-# Read the data
-data = pd.read_csv("../data/processed/processed_df.csv", index_col=0)
-
-# App server
-app = Dash(
-    __name__,
-    title="Vancouver Crime Dashboard",
-    external_stylesheets=[dbc.themes.BOOTSTRAP],
-)
-server = app.server
-
-"""Options"""
-# Options for neighbourhood
-opt_dropdown_neighbourhood = [
-    {"label": neighbourhood, "value": neighbourhood}
-    for neighbourhood in data["Neighborhood"].dropna().unique()
-]
-opt_dropdown_neighbourhood.remove(opt_dropdown_neighbourhood[-1])
-opt_dropdown_neighbourhood.remove(opt_dropdown_neighbourhood[-4])
-
-# Options for time
-opt_dropdown_time = [
-    {"label": "Day", "value": "Day"},
-    {"label": "Night", "value": "Night"},
-    {"label": "Day and Night", "value": "Day and Night"},
-]
-
-# Options for year
-opt_radio_year = [
-    {"label": "2021", "value": 2021},
-    {"label": "2020", "value": 2020},
-    {"label": "2019", "value": 2019},
-    {"label": "2018", "value": 2018},   
-    {"label": "2017", "value": 2017},
-]
-
-"""Cards"""
-# Summary card
-card1 = dbc.Card(
-    [
-        # Summary statistics
-        html.H4(
-            "Total Number of Crimes", className="card-title", style={"marginLeft": 50}
-        ),
-        html.Div(
-            id="summary", style={"color": "#E33B18", "fontSize": 25, "marginLeft": 140}
-        ),
-    ],
-    style={"width": "25rem", "marginLeft": 20},
-    body=True,
-    color="light",
-)
-
-# Filters card
-card2 = dbc.Card(
-    [
-        # Dropdown for neighbourhood
-        html.H5("Neighbourhood", className="text-dark"),
-        dcc.Dropdown(
-            id="neighbourhood_input",
-            value=["Kitsilano"],
-            options=opt_dropdown_neighbourhood,
-            className="dropdown",
-            multi=True
-        ),
-        html.Br(),
-        html.Br(),
-        html.Br(),
-        # Radio button for year
-        html.H5("Year", className="text-dark"),
-        dcc.RadioItems(
-            id="year_radio",
-            value=2021,
-            options=opt_radio_year,
-            className="radiobutton",
-            labelStyle={"display": "in-block", "marginLeft": 20},
-        ),
-        html.Br(),
-        html.Br(),
-        html.Br(),
-        # Dropdown for time
-        html.H5("Time", className="text-dark"),
-        dcc.Dropdown(
-            id="time_input",
-            value="Day and Night",
-            options=opt_dropdown_time,
-            className="dropdown",
-        )
-    ],
-    style={"width": "25rem", "marginLeft": 20},
-    body=True,
-    color="light",
-)
-
-# information
-card3 = dbc.Card(
-    [
-        html.H5("Information", className="text-dark"),
-        "Data used in this dashboard is sourced from the Vancouver Police Department at https://geodash.vpd.ca/opendata/. It has been filtered to only include incidents with location data from 2017 to 2021."
-        ],
-    style={"width": "25rem", "marginLeft": 20},
-    body=True,
-    color="light",
-        )
-
-"""Layouts"""
-# Filter layout
-filter_panel = [
-    html.H2("Vancouver Crime Dashboard", style={"marginLeft": 20}),
-    html.Br(),
-    html.Br(),
-    card1,
-    html.Br(),
-    html.Br(),
-    html.H4("Filters", style={"marginLeft": 20}),
-    card2,
-    html.Br(),
-    html.Br(),
-    card3
-]
-
-# Plots layout
-plot_body = [
-    dbc.Row(
-        [
-            dbc.Col(
-                [
-                    html.Iframe(
-                        id="bar_plot",
-                        className="bar_plot",
-                        style={
-                            "border-width": "0",
-                            "width": "100%",
-                            "height": "400px",
-                        },
-                    )
-                ],
-            ),
-            dbc.Col(
-                [
-                    dcc.Graph(id='map_plot',
-                              style={
-                                "border-width": "0",
-                                "width": "100%",
-                                "height": "400px",
-                            }
-                        )
-                  #  html.Iframe(
-                   #     id="map_plot",
-                   #     className="map_plot",
-                   #     style={
-                   #         "border-width": "0",
-                   #         "width": "100%",
-                   #         "height": "400px",
-                   #     },
-                   # )
-                ],
-            ),
-        ],
-    ),
-    html.Br(),
-    dbc.Row(
-        [
-            dbc.Col(
-                [
-                    html.Iframe(
-                        id="line_plot",
-                        className="line_plot",
-                        style={
-                            "border-width": "0",
-                            "width": "100%",
-                            "height": "400px",
-                        },
-                    )
-                ]
-            )
-        ]
-    ),
-]
-
-# Page layout
-page_layout = html.Div(
-    className="page_layout",
-    children=[
-        dbc.Row([html.Br()]),
-        dbc.Row(
-            [
-                dbc.Col(filter_panel, className="panel", width=3),
-                dbc.Col(plot_body, className="body"),
-            ]
-        ),
-    ],
-)
-
-# Overall layout
-app.layout = html.Div(id="main", className="app", children=page_layout)
-
-
-@app.callback(
-    Output("map_plot", "figure"),
-    Input("year_radio", "value"))
-def display_choropleth(year_radio):
-    df = pd.read_csv(r"../data/processed/map_df.csv")#px.data.election() # replace with your own data source
-    df = df[df['year']==int(year_radio)]
-    geojson = gpd.read_file("../data/vancouver.geojson")#px.data.election_geojson()
-    fig = px.choropleth(
-        df, geojson=geojson, color= df['count'],
-        locations="name", featureidkey="properties.name",
-        projection="mercator", #range_color=[0, df.count],
-        color_continuous_scale='YlOrRd',
-       # title = 'Crimes by Neighbourhood'
-        )
-    fig.update_geos(fitbounds="locations", visible=False)
-    fig.update_layout(margin={"r":20,"t":30,"l":0,"b":0})
-    return fig
-
-@app.callback(
-    Output("line_plot", "srcDoc"),
-    Input("time_input", "value"),
-    Input("neighbourhood_input", "value"),
-)
-def lineplot(time, neighbourhood):
-    data = pd.read_csv("../data/processed/processed_df.csv", index_col=0)
-    data = data[data.Neighborhood.isin(neighbourhood)]
-
-    if time == "Day":
-        data = data[data["TIME"] == "day"]
-
-    elif time == "Night":
-        data = data[data["TIME"] == "night"]
-
-    line_plot = (
-        alt.Chart(data, title="Crimes over Time")
-        .mark_line()
-        .encode(
-            x=alt.X("YEAR:O", title="Year"),
-            y=alt.Y("count(HOUR)", title="Number of Crimes"),
-            color=alt.Color(
-                "TIME", scale=alt.Scale(scheme="yelloworangered"), title="Time"
-            ),
-        )
-        .configure_axis(labelFontSize=14, titleFontSize=16)
-        .configure_legend(
-            titleFontSize=16,
-            labelFontSize=14,
-        )
-        .configure_title(fontSize=20)
-        .properties(width=1000, height=300)
-    )
-
-    return line_plot.to_html()
-
-
-@app.callback(
-    Output("bar_plot", "srcDoc"),
-    Input("neighbourhood_input", "value"),
-    Input("year_radio", "value"),
-)
-def barchart(neighbourhood, year):
-    data = pd.read_csv("../data/processed/processed_df.csv", index_col=0)
-    data = data[data["YEAR"] == year]
-    data = data[data.Neighborhood.isin(neighbourhood)]
-    data = pd.DataFrame(
-        data=data[["YEAR", "Type"]].value_counts(), columns=["Counts"]
-    ).reset_index(level=["YEAR", "Type"])
-
-    barchart = (
-        alt.Chart(data, title="Crimes by Type")
-        .mark_bar()
-        .encode(
-            x=alt.X(
-                "Type", sort="-y", axis=alt.Axis(labels=False), title="Type of Crime"
-            ),
-            y=alt.Y("Counts", title="Number of Crimes"),
-            color=alt.Color(
-                "Type", scale=alt.Scale(scheme="yelloworangered"), title="Type"
-            ),
-            tooltip=alt.Tooltip(["Type", "Counts"]),
-        )
-        .transform_window(
-            rank="rank(COUNTS)", sort=[alt.SortField("COUNTS", order="descending")]
-        )
-        .transform_filter((alt.datum.rank < 15))
-        .configure_axis(labelFontSize=14, titleFontSize=16)
-        .configure_legend(
-            titleFontSize=16,
-            labelFontSize=14,
-        )
-        .configure_title(fontSize=20)
-        .properties(width=300, height=300)
-    )
-    return barchart.to_html()
-
-
-@app.callback(
-    Output("summary", "children"),
-    Input("neighbourhood_input", "value"),
-    Input("year_radio", "value"),
-)
-def summary(neighbourhood, year):
-    data = pd.read_csv("../data/processed/processed_df.csv", index_col=0)
-    data = data[data["YEAR"] == year]
-    data = data[data.Neighborhood.isin(neighbourhood)]
-    return len(data)
-
-
-if __name__ == "__main__":
-    app.run_server(debug=True)
